+++ simple-stack/src/main/java/com/zhuinden/simplestack/BackstackManager.java
@@ -337,6 +337,7 @@ public class BackstackManager
         for(String scope : scopes) {
             scopeManager.destroyScope(scope);
         }
+        scopeManager.finalizeScopes();
     }
 
     /**